pyproject.toml
[tool.poetry]
name = "sendcloud"
version = "0.1.0"
description = ""
authors = ["Wouter Lansu <wfrlansu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
django-ninja = "^1.2.2"
celery = "^5.4.0"
redis = "^5.0.7"
httpx = "^0.27.0"
django-stubs-ext = "^5.0.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
pytest-django = "^4.8.0"
mypy = "^1.11.1"
ruff = "^0.5.7"
django-stubs = { version = "^5.0.4", extras = [ "compatible-mypy" ] }
celery-stubs = "^0.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "sendcloud.settings"
